+++ src/Test/Integration/Model/Payment/FactoryTest.php
@@ -20,6 +20,8 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
 
+declare(strict_types=1);
+
 namespace Shopgate\Import\Test\Integration\Model\Payment;
 
 use Magento\Framework\Exception\LocalizedException;
@@ -54,8 +56,10 @@ class FactoryTest extends TestCase
      * @dataProvider paymentMethodProvider
      *
      * @throws LocalizedException
+     * @throws \ShopgateLibraryException
      *
      * @magentoConfigFixture current_store payment/braintree/active 1
+     * @magentoConfigFixture current_store payment/braintree_paypal/active 1
      */
     public function testPaymentMethodMapping($methodCode, $expectedPaymentMethod): void
     {
@@ -77,6 +81,10 @@ class FactoryTest extends TestCase
             'Cash on delivery'                          => ['COD', 'cashondelivery'],
             'Invoice'                                   => ['INVOICE', 'checkmo'],
             'Braintree Credit Card'                     => ['BRAINTR_CC', 'braintree'],
+            'Braintree PayPal'                          => ['BRAINTR_PP', 'braintree_paypal'],
+            'return shopgate as default payment method' => ['PREPAY', 'shopgate'],
+            'Braintree Credit Card'                     => ['BRAINTR_CC', 'braintree'],
+            'Braintree PayPal'                          => ['BRAINTR_PP', 'braintree_paypal'],
         ];
     }
 }